+++ restfull-api/middleware/async-try-catch.js
@@ -1,0 +1,9 @@
+module.export = function (routeHandler) {
+  return async (req, res, next) => {
+    try {
+      await routeHandler(req, res);
+    } catch (err) {
+      next(err);
+    }
+  };
+};
