+++ restfull-api/middleware/error.js
@@ -1,0 +1,3 @@
+module.exports = function (err, req, res, next) {
+  res.status(500).json({ message: err.message });
+};
