+++ restfull-api/server.js
@@ -1,5 +1,6 @@
 require("dotenv").config();
 
+const error = require("./middleware/error");
 const express = require("express");
 const app = express();
 const mongoose = require("mongoose");
@@ -10,6 +11,7 @@ mongoose.connection.on("error", (error) => console.error(error.text));
 mongoose.connection.once("open", () => console.log("Connected to Database..."));
 
 app.use(express.json());
+app.use(error);
 
 const tdosRouter = require("./routes/todos");
 app.use("/api/todos", tdosRouter);
